refactor(storage): share one MergingIterator across the chunked backends
The chunked-snapshot work left redb, fjall, memory and rocksdb each
holding a 221-line copy of the same merging iterator, differing only in
which backend their doc comments named. Half the change already shared
a module; this makes the other half match.

lark keeps its own iterator: it was already lazy and merges differently.
