+++ crates/storage/src/backends/fjall/mod.rs
@@ -1,6 +1,5 @@
 pub mod config;
 mod errors;
-mod iterator;
 mod store;
 mod transaction;
