+++ crates/storage/src/backends/rocksdb/mod.rs
@@ -1,6 +1,5 @@
 pub mod config;
 mod errors;
-mod iterator;
 mod metrics;
 mod store;
 mod transaction;
